ra: move login handling out of settings
Move RetroAchievements login and HTTP request handling from
the settings panel into a new desktop-ui RetroAchievements service.

This keeps the settings panel focused on UI state and credential input,
while the service owns the rcheevos login request, libcurl transport,
response parsing, and disabled-build fallback.
