+++ desktop-ui/cmake/sources.cmake
@@ -24,6 +24,8 @@ target_sources(
 
 target_sources(desktop-ui PRIVATE resource/resource.cpp resource/resource.hpp)
 
+target_sources(desktop-ui PRIVATE retroachievements/retroachievements.cpp retroachievements/retroachievements.hpp)
+
 target_sources(
   desktop-ui
   PRIVATE
